pyproject.toml
[tool.poetry]
name = "text-translator-project"
version = "0.1.0"
description = ""
authors = ["anderson89marques <andersonoanjo18@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.6"
djangorestframework = "^3.14.0"
requests = "^2.31.0"
drf-spectacular = "^0.26.5"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.9.1"
flake8 = "^6.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
